requirements.txt
chat_downloader==0.2.8
numpy==1.23.5
pandas==1.5.3
plotly==5.17.0
pytube==15.0.0
PyYAML==6.0
PyYAML==6.0.1
scanpy==1.10.2
scikit_learn==1.2.1
sentence_transformers==2.2.2
torch==2.3.1
umap==0.1.1
umap_learn==0.5.4
youtube_comment_downloader==0.1.76
